+++ tests/util/macros.cpp
