+++ tests/util/test_ofstream.cpp
