+++ tests/util/test_ofstream.hpp
